Add files via upload
This repository contains some basic and medium-level jupyter notebooks for solutions to assignments. These assignments were part of the summer analytics course organized by consulting and analytics club IITG. the datasets are attached to the repository. the text part of the notebooks is provided by the course mentors and codes are edited.